backend/requirements.txt
Django>=3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.2.1
psycopg2>=2.7,<3.0
